add guards to students and papers

+++ server/src/students/students.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { HttpException, Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateStudentDto } from './dto/create-student.dto';
 import { UpdateStudentDto } from './dto/update-student.dto';
@@ -7,20 +7,47 @@ import { UpdateStudentDto } from './dto/update-student.dto';
 export class StudentsService {
   constructor(private readonly prisma: PrismaService) {}
 
-  async getStudents(){
-    return await this.prisma.students.findMany({})
+  async getStudents(user_id: number){
+    return this.prisma.students.findMany({
+      where: {
+        groups: {
+          some: {
+            user_id: +user_id,
+          },
+        },
+      },
+    });
   }
 
-  async getStudent(id: number){
-    const student = await this.prisma.students.findUnique({where: {student_id: +id}})
+  async getStudent(id: number, user_id: number){
+    const student = await this.prisma.students.findUnique({
+      where: { student_id: +id },
+      include: {
+        groups: {
+          where: {
+            user_id: +user_id,
+          },
+        },
+      },
+    });
+
     if (!student) {
       throw new NotFoundException(`Can't find this student : ${id}`)
     }
+
+    if (!student.groups.some((group) => group.user_id === user_id)) {
+      throw new HttpException('Unauthorized', 401);
+    }
+  
     return student;
   }
 
-  async createStudent(studentDto:  CreateStudentDto){
-    const { groups, ...restDto } = studentDto;
+  async createStudent(studentDto:  CreateStudentDto, userId: number){
+    const { user_id, groups, ...restDto } = studentDto;
+
+    if(user_id != userId){
+      throw new HttpException('Unauthorized', 401);
+    }
 
     if (groups && Array.isArray(groups)) {
       const group_ids = groups.map(group_id => ({ group_id: group_id }));
@@ -37,10 +64,22 @@ export class StudentsService {
     }
   }
 
+  async updateStudent(id: number, studentDto: UpdateStudentDto, userId: number){
+    const { user_id, groups, ...restDto } = studentDto;
 
-  async updateStudent(id: number, studentDto: UpdateStudentDto){
-    const { groups, ...restDto } = studentDto;
-    
+    const existingStudent = await this.prisma.students.findUnique({
+      where: { student_id: +id },
+      include: { groups: true }, 
+    });
+  
+    if (!existingStudent) {
+      throw new NotFoundException(`Can't find this student: ${id}`);
+    }
+  
+    if (!existingStudent.groups.some((group) => group.user_id === userId)) {
+      throw new HttpException('Unauthorized', 401);
+    }
+
     if (groups && Array.isArray(groups)) {
       const group_ids = groups.map(group_id => ({ group_id: group_id }));
 
@@ -60,7 +99,26 @@ export class StudentsService {
     }
   }
 
-  async deleteStudent(id: number){
+  async deleteStudent(id: number, user_id: number){
+    const student = await this.prisma.students.findUnique({
+      where: { student_id: +id },
+      include: {
+        groups: {
+          where: {
+            user_id: +user_id,
+          },
+        },
+      },
+    });
+
+    if (!student) {
+      throw new NotFoundException(`Can't find this student : ${id}`)
+    }
+
+    if (!student.groups.some((group) => group.user_id === user_id)) {
+      throw new HttpException('Unauthorized', 401);
+    }
+
     return await this.prisma.students.delete({where: {student_id: +id}})
   }
